Rename index lab users fn
